# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityCategory } from "./activityCategory";

export interface Activity {
  id: string;
  name: string;
  category: ActivityCategory;
  /** Duration in minutes */
  estimated_duration: number;
  description: string;
  address?: string;
  lat?: number;
  lng?: number;
  score?: number;
}