# lib/api-zod/src/generated/types/activityCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ActivityCategory =
  (typeof ActivityCategory)[keyof typeof ActivityCategory];

export const ActivityCategory = {
  museum: "museum",
  landmark: "landmark",
  park: "park",
  restaurant: "restaurant",
  shopping: "shopping",
  entertainment: "entertainment",
  cultural: "cultural",
  outdoor: "outdoor",
  nightlife: "nightlife",
  tour: "tour",
} as const;