# lib/api-zod/src/generated/types/dayPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from "./activity";

export interface DayPlan {
  day: number;
  activities: Activity[];
}