# lib/api-zod/src/generated/types/generateItineraryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TravelPace } from "./travelPace";
import type { UserInterest } from "./userInterest";

export interface GenerateItineraryRequest {
  city: string;
  /**
   * @minimum 1
   * @maximum 3
   */
  tripDays: number;
  /** @maxItems 3 */
  interests: UserInterest[];
  pace: TravelPace;
  /** Optional seed for reproducible results */
  seed?: number;
}