# lib/api-zod/src/generated/types/generateItineraryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DayPlan } from "./dayPlan";
import type { TravelPace } from "./travelPace";
import type { UserInterest } from "./userInterest";

export interface GenerateItineraryResponse {
  city: string;
  tripDays: number;
  pace: TravelPace;
  interests: UserInterest[];
  days: DayPlan[];
  isOptimized: boolean;
  fallbackTier?: number | null;
  generatedAt: string;
}