# lib/api-zod/src/generated/types/savedItinerary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DayPlan } from "./dayPlan";
import type { TravelPace } from "./travelPace";
import type { UserInterest } from "./userInterest";

export interface SavedItinerary {
  id: string;
  userId: string;
  city: string;
  tripDays: number;
  pace: TravelPace;
  interests: UserInterest[];
  days: DayPlan[];
  isOptimized: boolean;
  createdAt: string;
  updatedAt: string;
}