# lib/api-zod/src/generated/types/travelPace.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TravelPace = (typeof TravelPace)[keyof typeof TravelPace];

export const TravelPace = {
  relaxed: "relaxed",
  standard: "standard",
  packed: "packed",
} as const;