# lib/api-zod/src/generated/types/userInterest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type UserInterest = (typeof UserInterest)[keyof typeof UserInterest];

export const UserInterest = {
  history: "history",
  art: "art",
  food: "food",
  nature: "nature",
  architecture: "architecture",
  nightlife: "nightlife",
  shopping: "shopping",
  sports: "sports",
} as const;